Switched to using a different method for generating the script used as a entry point into the tool.
The previous method works fine on Linux but on Windows we get a script
that cannot be executed directly from the command line (we have to
prepend the python executable). This new method generates an entry point
script on Linux and on windows it actually generates a .exe file as
entry point so the tool can be run directly from the command line :)

I've switched from distutils.core to setuptools to squash the warning I
was getting about entry_points being ignored. Despite that it was still
working correctly on distutils.

Note I've still kept bin/OutputCheck because the testing framework still
depends on it and its nice to have a script I can run directly in the
repo.